Update: NULL check in is_full from == to !=
Previously, the if statement in is_full was 'if (queue == NULL ||
queue->count == queue->size)'. This meant that even when queue was a
NULL pointer, the function would return 1 indicating that it was full.
This shouldn't be the case. Therefore, the statement was changed to 'if
(queue != NULL && queue->count == queue->size)'